Better include structure, put Legacy stuff to Legacy headers
