+++ CMakeLists.txt
@@ -404,6 +404,7 @@ set(CORE_SOURCE
         src/ParticlePatches.cpp
         src/ParticleSpecies.cpp
         src/Record.cpp
+        src/ReadIterations.cpp
         src/RecordComponent.cpp
         src/Series.cpp
         src/UnitDimension.cpp
@@ -422,6 +423,7 @@ set(CORE_SOURCE
         src/helper/list_series.cpp
         src/snapshots/ContainerImpls.cpp
         src/snapshots/ContainerTraits.cpp
+        src/snapshots/IteratorHelpers.cpp
         src/snapshots/IteratorTraits.cpp
         src/snapshots/RandomAccessIterator.cpp
         src/snapshots/Snapshots.cpp
