+++ src/snapshots/StatefulIterator.cpp
@@ -966,67 +966,4 @@ StatefulIterator::operator bool() const
 {
     return m_data->has_value();
 }
-
-LegacyIteratorAdaptor::LegacyIteratorAdaptor(StatefulIterator iterator)
-    : m_iterator(std::move(iterator))
-{}
-
-auto LegacyIteratorAdaptor::operator*() const -> value_type
-{
-    return m_iterator.operator*();
-}
-
-auto LegacyIteratorAdaptor::operator++() -> LegacyIteratorAdaptor &
-{
-    ++m_iterator;
-    return *this;
-}
-
-auto LegacyIteratorAdaptor::end() -> LegacyIteratorAdaptor
-{
-    return StatefulIterator::end();
-}
-
-auto LegacyIteratorAdaptor::operator==(LegacyIteratorAdaptor const &other) const
-    -> bool
-{
-    return m_iterator == other.m_iterator;
-}
-
-auto LegacyIteratorAdaptor::operator!=(LegacyIteratorAdaptor const &other) const
-    -> bool
-{
-    return m_iterator != other.m_iterator;
-}
-
-ReadIterations::ReadIterations(
-    Series series,
-    Access access,
-    std::optional<internal::ParsePreference> parsePreference)
-    : m_series(std::move(series)), m_parsePreference(parsePreference)
-{
-    auto &data = m_series.get();
-    if (access == Access::READ_LINEAR && !data.m_sharedReadIterations)
-    {
-        // Open the iterator now already, so that metadata may already be read
-        data.m_sharedReadIterations = std::make_unique<StatefulIterator>(
-            StatefulIterator::tag_read, m_series, m_parsePreference);
-    }
-}
-
-ReadIterations::iterator_t ReadIterations::begin()
-{
-    auto &series = m_series.get();
-    if (!series.m_sharedReadIterations)
-    {
-        series.m_sharedReadIterations = std::make_unique<StatefulIterator>(
-            StatefulIterator::tag_read, m_series, m_parsePreference);
-    }
-    return *series.m_sharedReadIterations;
-}
-
-ReadIterations::iterator_t ReadIterations::end()
-{
-    return iterator_t::end();
-}
 } // namespace openPMD
